Cleanups and change to debug

+++ src/stores/PouchStore.js
@@ -1,41 +1,43 @@
-/* vim :set ts=2 sw=2 sts=2 et : */
-"use strict";
-var alt = require('../alt');
+'use strict';
+
 var merge = require('object-assign');
+var debug = require('react-debug');
 var PouchDB = require('pouchdb');
 
 var PouchActions = require('../actions/PouchActions');
 
+window.localStorage.debug = '*';
+
 class PouchStore {
   constructor(name, view, key, readyFunc) {
-    console.log('constructor', arguments);
+    debug(this, 'constructor', arguments);
     this.bindActions(PouchActions);
     this.docs = {};
     this.db = new PouchDB(name);
     this.name = name;
-    console.log("db", this.db);
-    console.log("name", this.name);
+    debug(this, 'db', this.db);
+    debug(this, 'name', this.name);
     var options = {
       since: 'now',
       live: true,
       include_docs: true
     };
     var watchChanges = function(result) {
-      console.log('result', result);
+      debug(this, 'result', result);
       if(result && result.rows) {
-        var newrows = result.rows.map(function(row){return row.doc});
+        var newrows = result.rows.map(function(row){return row.doc;});
         this.onUpdateAll(newrows);
       }
       this.changes = this.db.changes(options).on('change', function(change) {
-        console.log('changes change', change);
+        debug(this, 'changes change', change);
         var doc = change.doc;
-        if(doc) { 
+        if(doc) {
           this.onUpdateAll([doc]);
         }
       }.bind(this)).on('complete', function(info) {
-        console.log('changes complete', info);
+        debug(this, 'changes complete', info);
       }.bind(this)).on('error', function (err) {
-        console.log('changes error', info);
+        debug(this, 'changes error', err);
       }.bind(this));
     }.bind(this);
     if(view) {
@@ -45,7 +47,7 @@ class PouchStore {
     } else {
       this.db.allDocs({include_docs: true}).then(watchChanges);
     }
-    if(key) { 
+    if(key) {
       this.key = key;
     } else {
       this.key = '_id';
@@ -56,55 +58,65 @@ class PouchStore {
   }
 
   onPut(doc) {
-    console.log('put', doc);
-    this.db.put(doc).then(function(result) {
-        console.log('put result', result);
-    }).catch(function(err) {
-        console.log('put error: ', err);
-    });
+    debug(this, 'put', doc);
+    if(this.db) {
+      this.db.put(doc).then(function(result) {
+          debug(this, 'put result', result);
+      }).catch(function(err) {
+          debug(this, 'put error: ', err);
+      });
+    }
   }
 
   onUpdateAll(docs) {
-    console.log('update', docs);
-    if(Array.isArray(docs)){
-      for(var i = 0; i < docs.length; i++) {
-        var doc = docs[i];
-        console.log('doc', doc);
-        var key = doc[this.key];
-        console.log('key', key);
-        if(doc._deleted && key in this.docs) {
-          delete this.docs[key];
-        } else {
-          this.docs[key] = doc;
+    debug(this, 'update', docs);
+    if(this.db) {
+      if(Array.isArray(docs)){
+        for(var i = 0; i < docs.length; i++) {
+          var doc = docs[i];
+          debug(this, 'doc', doc);
+          var key = doc[this.key];
+          debug(this, 'key', key);
+          if(doc._deleted && key in this.docs) {
+            delete this.docs[key];
+          } else {
+            this.docs[key] = doc;
+          }
+        }
+      } else {
+        for(var key2 in this.docs) {
+          this.docs[key2] = merge(this.docs[key2], docs);
         }
       }
-    } else {
-      for(var key in this.docs) {
-        this.docs[key] = merge(this.docs[key], docs);
-      }
+      debug(this, 'docs', this.docs);
+      this.emitChange();
     }
-    console.log('docs', this.docs);
-    this.emitChange();
   }
 
   onRemove(doc) {
-    console.log('remove', doc);
-    this.db.remove(doc).then(function(result) {
-        console.log('remove result', result);
-    }).catch(function(err) {
-        console.log('remove error: ', err);
-    });
+    if(this.db) {
+      debug(this, 'remove', doc);
+      this.db.remove(doc).then(function(result) {
+          debug(this, 'remove result', result);
+      }).catch(function(err) {
+          debug(this, 'remove error: ', err);
+      });
+    }
   }
 
   onSync(destination) {
-    console.log('sync', this.name, destination);
-    PouchDB.sync(this.name, destination);
+    debug(this, 'sync', this.name, destination);
+    if(this.db) {
+      PouchDB.sync(this.name, destination);
+    }
   }
 
   onCreateDb(name) {
+    debug(this, 'createDB', name);
   }
 
   onDeleteDb(name) {
+    debug(this, 'deleteDB', name);
   }
 }
